+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 1.0.2 (COMING)
 * DisplayMetricsUtil added
+* RecyclerViewUtils.setItemViewSwipeAlphaValue added
 
 ## Version 1.0.1 (2016-05-12)
 * Very first release